Use int instead of long

+++ cupy/random/generator.py
@@ -82,7 +82,7 @@ class RandomState(object):
         if seed is None:
             try:
                 seed_str = os.urandom(8)
-                seed = numpy.uint64(long(seed_str.encode('hex'), 16))
+                seed = numpy.uint64(int(seed_str.encode('hex'), 16))
             except NotImplementedError:
                 seed = numpy.uint64(time.clock() * 1000000)
         else:
